Good first try!
I commented out the title you had and changed it to a page title.  When you save the page to a file on your computer, you will call the home page index.html.  The other pages can be named anything you want, but remember to keep it short and something that makes sense to what the page is about like about.html or contact.html.  The <title> is what shows up in the tab when the page is opened in the browser.
The file name should always end in .html .